Refactor: Restructure project with package refactoring
This commit restructures the project by moving classes into more specific packages.
- `OnboardingActivity`, `LoginActivity`, `ForgotPasswordActivity`, and `SignupActivity` have been moved to the `authentication` package.
- `NotificationHelper`, `NotificationCenterActivity`, and `FridgeMonitoringService`, `DatabaseHelper` have been moved to the `notifications` package.
- `FridgeConditionsActivity` has been moved to the `fridge_conditions` package.
- `Product`, `BarcodeScannerActivity`, `InventoryActivity`, `InventoryAdapter`, `MergeSort`, `DateInfo`, `AddProductDialogFragment`, `AddManualProductDialogFragment` and `CategoryUtils` have been moved to the `inventory` package.
- `DashboardActivity` and `ProfileActivity` have been moved to the `dashboard` package.
- `SettingsActivity` has been moved to the `settings` package.
- `ShoppingListActivity` has been moved to the `shopping_list` package.
- Updated class name FridgeConditions to FridgeConditionsActivity
- Updated all the references to the classes due to package renaming
- Updated all the references to layout,string and color resources
