+++ app/src/main/AndroidManifest.xml
@@ -20,41 +20,41 @@
         android:theme="@style/Theme.SmartFoodInventoryTracker"
         tools:targetApi="31">
         <activity
-            android:name=".SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".NotificationCenter"
+            android:name=".notifications.NotificationCenterActivity"
             android:exported="false" />
         <activity
-            android:name=".ProfileActivity"
+            android:name=".dashboard.ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".FridgeConditions"
+            android:name=".fridge_conditions.FridgeConditionsActivity"
             android:exported="false" />
         <activity
-            android:name=".ShoppingList"
+            android:name=".shopping_list.ShoppingListActivity"
             android:exported="false" />
         <activity
-            android:name=".InventoryActivity"
+            android:name=".inventory.InventoryActivity"
             android:exported="false" /> <!-- ✅ Dashboard Activity -->
         <activity
-            android:name=".DashboardActivity"
+            android:name=".dashboard.DashboardActivity"
             android:exported="false" /> <!-- ✅ Forgot Password Activity -->
         <activity
-            android:name=".ForgotPasswordActivity"
+            android:name=".authentication.ForgotPasswordActivity"
             android:exported="false" /> <!-- ✅ Signup Activity -->
         <activity
-            android:name=".SignupActivity"
+            android:name=".authentication.SignupActivity"
             android:exported="false" /> <!-- ✅ Login Activity as Launcher -->
         <activity
-            android:name=".LoginActivity"
+            android:name=".authentication.LoginActivity"
             android:exported="false" /> <!-- ✅ Main Activity (Unused or Future Use) -->
         <activity
             android:name=".MainActivity"
             android:exported="false" />
-        <activity android:name=".BarcodeScannerActivity" />
+        <activity android:name=".inventory.BarcodeScannerActivity" />
         <activity
-            android:name=".OnboardingActivity"
+            android:name=".authentication.OnboardingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
