+++ app/src/main/java/com/example/smartfoodinventorytracker/dashboard/ProfileActivity.java
@@ -1,12 +1,12 @@
-package com.example.smartfoodinventorytracker;
+package com.example.smartfoodinventorytracker.dashboard;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.view.View;
 import android.widget.*;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.smartfoodinventorytracker.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
