+++ app/src/main/java/com/example/smartfoodinventorytracker/inventory/AddProductDialogFragment.java
@@ -1,17 +1,18 @@
-package com.example.smartfoodinventorytracker;
+package com.example.smartfoodinventorytracker.inventory;
 
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.Button;
 import android.widget.LinearLayout;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
+
+import com.example.smartfoodinventorytracker.R;
 
 public class AddProductDialogFragment extends DialogFragment {
