+++ app/src/main/java/com/example/smartfoodinventorytracker/inventory/DateInfo.java
@@ -1,13 +1,4 @@
-package com.example.smartfoodinventorytracker;
-
-import androidx.annotation.NonNull;
-
-import com.google.type.DateTime;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.List;
+package com.example.smartfoodinventorytracker.inventory;
 
 public class DateInfo {
