+++ app/src/main/res/layout/activity_dashboard.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:theme="@style/Theme.SmartFoodInventoryTracker"
-    tools:context=".DashboardActivity">
+    tools:context=".dashboard.DashboardActivity">
 
     <!-- Main Content -->
     <androidx.constraintlayout.widget.ConstraintLayout
