+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgotPasswordActivity"
+    tools:context=".authentication.ForgotPasswordActivity"
     android:theme="@style/Theme.SmartFoodInventoryTracker"
     android:padding="20dp">
