+++ app/src/main/res/layout/activity_fridge_conditions.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FridgeConditions">
+    tools:context=".fridge_conditions.FridgeConditionsActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
