+++ app/src/main/res/layout/activity_notification_center.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:theme="@style/Theme.SmartFoodInventoryTracker"
     android:layout_height="match_parent"
-    tools:context=".NotificationCenter">
+    tools:context=".notifications.NotificationCenterActivity">
 
     <!-- Toolbar -->
     <androidx.appcompat.widget.Toolbar
