+++ app/src/main/res/layout/activity_shopping_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:theme="@style/Theme.SmartFoodInventoryTracker"
     android:layout_height="match_parent"
-    tools:context=".ShoppingList">
+    tools:context=".shopping_list.ShoppingListActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
