+++ app/src/main/res/layout/activity_signup.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:theme="@style/Theme.SmartFoodInventoryTracker"
     android:layout_height="match_parent"
-    tools:context=".SignupActivity"
+    tools:context=".authentication.SignupActivity"
     android:padding="20dp">
 
     <!-- App Logo -->
